+++ features/step_definitions/mailer_step.rb
@@ -27,13 +27,23 @@ Given(/^the sender is set to "([^"]*)"$/) do |sender|
   @options["email_sender"] = sender
 end
 
+Given(/^the ExactTarget header is set$/) do
+  @et_header = "something-very-long"
+  @options[:et_route_key] = @et_header
+end
+
+Given(/^the message id header is set$/) do
+  @message_id = "message_id"
+  @options[:messageId] = @message_id
+end
+
 When(/^I do not provide the template variable "([^"]*)"$/) do |template_variable|
   # The template_variable will either be in the root of the hash or inside the templateVariables sub hash
   @options["templateVariables"].delete template_variable
 end
 
 When(/^the message is rejected$/) do
-  delivery_id = $nacked.pop
+  delivery_id = $rejected.pop
   expect(delivery_id ).to eq @delivery_id
 end
 
@@ -41,7 +51,11 @@ When(/^the message is processed$/) do
   ActionMailer::Base.delivery_method = :test
   fake_delivery_options = Bunny::DeliveryInfo.new
   @delivery_id = fake_delivery_options.identifier
-  $mailer_daemon.process_mail(fake_delivery_options, @options)
+  begin
+    $mailer_daemon.process_message(fake_delivery_options, @options)
+  rescue Exception => e
+    @process_failure_exception = e
+  end
 end
 
 Then(/^an email is delivered to "(.*)"$/) do |email_address|
@@ -74,11 +88,19 @@ Then(/^I do not deliver an email to "(.*)"$/) do |email|
 end
 
 Then(/^I get a message sent the "([^"]*)" queue$/) do |queue_suffix|
-  deliver_id = $nacked.pop
+  deliver_id = $rejected.pop
   expect(deliver_id).to eq @delivery_id
 end
 
 Then(/^the sender is "([^"]*)"$/) do |sender|
   # There seems to be no way of getting the sender's name from the Mail::Message object as of actionmailer 4.0.0
   expect(@email.from[0]).to eq sender
+end
+
+Then(/^it has the exact target headers$/) do
+  expect(@email.header["x-et-route"].to_s).to eq(@et_header)
+end
+
+Then(/^it has the message id header$/) do
+  expect(@email.header["X-BBB-Message-Id"].to_s).to eq(@message_id)
 end